requirements.txt
youtube-search==2.1.2
youtube-transcript-api==0.6.3
pandas==2.2.3
google-generativeai==0.8.4
google-cloud-texttospeech==2.25.1
google-cloud-storage==2.19.0
python-dotenv==1.0.1
pydub==0.25.1
streamlit==1.41.1
gtts==2.5.4
tabulate==0.9.0
IMDBpy==2022.7.9